# contracts/core-business.openapi.yaml
openapi: 3.1.0
...
info:
  title: Smart Campus Core Business Policy API
  version: 1.0.0
  summary: Realtime access-policy evaluation for the Smart Campus platform.
  description: |
    Contract negotiated between Access Gate (consumer) and Core Business
    (provider). The API evaluates access requests, manages policies, and
    exposes decisions and alerts for audit.
  contact:
    name: Core Business Team 6A
    email: trongnamm2k5@gmail.com
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:4010
    description: Prism mock server
  - url: http://localhost:8000
    description: Local Core Business service
...
paths:
...
  /api/v1/access-events:
    post:
      tags: [Integration]
      summary: Process an Access Gate event
      description: Evaluates realtime access and forwards the result to Analytics and Notification when required.
      operationId: ingestAccessEvent
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessEvaluationRequest"
      responses:
        "200":
          description: Access event processed synchronously.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/sensor-events:
    post:
      tags: [Integration]
      summary: Process an IoT sensor event
      description: Applies sensor policy and forwards the normalized result to partner services.
      operationId: ingestSensorEvent
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorEvaluationRequest"
      responses:
        "202":
          description: Sensor event accepted and processed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/detections:
    post:
      tags: [Integration]
      summary: Process an AI Vision detection
      description: Applies detection policy and forwards security outcomes to partner services.
      operationId: ingestDetection
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DetectionEvaluationRequest"
      responses:
        "202":
          description: Detection accepted and processed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /policies/evaluate-access:
    post:
      tags: [Policies]
      summary: Evaluate a realtime access request
      description: |
        Evaluates an Access Gate request against the active policy. Access Gate
        must fail closed when this operation times out or returns a 5xx response.
      operationId: evaluateAccessPolicy
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessEvaluationRequest"
            examples:
              employeeEntry:
                value:
                  requestId: 1b8a6f7f-f27e-4b98-b884-22fb705ec587
                  cardId: CARD-000123
                  gateId: GATE-A1
                  direction: ENTRY
                  occurredAt: "2026-05-19T08:15:00Z"
                  subject:
                    subjectId: EMP-0007
                    role: STAFF
                    cardStatus: ACTIVE
                    zone: ADMIN
      responses:
        "200":
          description: Policy decision returned.
          headers:
            X-Correlation-Id:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessDecision"
              examples:
                allowed:
                  value:
                    decisionId: c439b6fe-7144-47aa-ab60-65d17ebedb78
                    requestId: 1b8a6f7f-f27e-4b98-b884-22fb705ec587
                    decision: ALLOW
                    reasonCode: POLICY_ALLOW
                    policyId: 7e34be8b-1da8-483e-b5ae-28f8662d0ac7
                    evaluatedAt: "2026-05-19T08:15:00Z"
                    expiresAt: "2026-05-19T08:15:30Z"
                    correlationId: 80cddcc0-72e8-41da-8425-22a263f11254
                    explanation: Active staff card is allowed in ADMIN zone.
                denied:
                  value:
                    decisionId: 4e7c8a4d-b530-4b2b-8f7f-e34ab3c94fa4
                    requestId: 1b8a6f7f-f27e-4b98-b884-22fb705ec587
                    decision: DENY
                    reasonCode: POLICY_DENY_REVOKED
                    policyId: 7e34be8b-1da8-483e-b5ae-28f8662d0ac7
                    evaluatedAt: "2026-05-19T08:15:00Z"
                    expiresAt: "2026-05-19T08:15:30Z"
                    correlationId: 80cddcc0-72e8-41da-8425-22a263f11254
                    explanation: Card status is REVOKED.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /policies/evaluate-sensor:
    post:
      tags: [Policies]
      summary: Evaluate a sensor reading
      description: Applies threshold policy to IoT telemetry and may create an alert.
      operationId: evaluateSensorPolicy
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorEvaluationRequest"
            examples:
              highTemperature:
                value:
                  requestId: c95458d1-aa71-444f-82ad-a9d45337c12b
                  deviceId: SENSOR-A101
                  metric: TEMPERATURE
                  value: 42.5
                  unit: CELSIUS
                  occurredAt: "2026-05-19T08:20:00Z"
      responses:
        "200":
          description: Sensor policy outcome.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorDecision"
              examples:
                warning:
                  value:
                    decisionId: bb8c175d-5a5a-42ea-a29a-f7639130f69b
                    requestId: c95458d1-aa71-444f-82ad-a9d45337c12b
                    outcome: ALERT
                    reasonCode: SENSOR_THRESHOLD_CRITICAL
                    policyId: 98ae19f6-13f6-4cc5-aa8f-3b76cb8c68ec
                    alertId: 6644f568-85df-4913-b959-50df1616d58b
                    evaluatedAt: "2026-05-19T08:20:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /policies/evaluate-detection:
    post:
      tags: [Policies]
      summary: Evaluate an AI Vision detection
      description: Applies security policy to a normalized AI Vision result.
      operationId: evaluateDetectionPolicy
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DetectionEvaluationRequest"
            examples:
              unknownPerson:
                value:
                  requestId: 8cc9bb0d-ee85-41e5-a3fa-f4155c7286d8
                  detectionId: 4ec96900-37e0-47eb-9021-b95885c089d8
                  cameraId: CAMERA-LOBBY-01
                  label: UNKNOWN_PERSON
                  confidence: 0.96
                  occurredAt: "2026-05-19T08:21:00Z"
      responses:
        "200":
          description: Detection policy outcome.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetectionDecision"
              examples:
                review:
                  value:
                    decisionId: fa4ee2a2-58e8-4d2c-b41c-811013b8af56
                    requestId: 8cc9bb0d-ee85-41e5-a3fa-f4155c7286d8
                    outcome: ALERT
                    reasonCode: UNKNOWN_PERSON_HIGH_CONFIDENCE
                    policyId: 26a2f3c5-32ce-4c1d-b66e-d3e6a900989e
                    alertId: 289ef942-bd64-4e95-8645-87a30aa54505
                    evaluatedAt: "2026-05-19T08:21:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /decisions/{decisionId}:
    get:
      tags: [Decisions]
      summary: Get a policy decision
      description: Retrieves an immutable decision for incident investigation.
      operationId: getDecision
      parameters:
        - $ref: "#/components/parameters/DecisionId"
      responses:
        "200":
          description: Decision found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessDecision"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Unique key reused only when retrying the same request.
      schema:
        type: string
        minLength: 8
        maxLength: 128
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Cross-service trace identifier.
      schema:
        type: string
        format: uuid
...
  headers:
    CorrelationId:
      description: Correlation identifier assigned to the request.
      schema:
        type: string
        format: uuid
  responses:
...
    UnprocessableEntity:
      description: JSON is valid but violates a business rule.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    AccessEvaluationRequest:
      type: object
      additionalProperties: false
      required: [requestId, cardId, gateId, direction, occurredAt, subject]
      properties:
        requestId:
          type: string
          format: uuid
        cardId:
          type: string
          pattern: "^CARD-[A-Z0-9]{6,20}$"
        gateId:
          type: string
          pattern: "^GATE-[A-Z0-9-]{2,20}$"
        direction:
          type: string
          enum: [ENTRY, EXIT]
        occurredAt:
          type: string
          format: date-time
        subject:
          $ref: "#/components/schemas/AccessSubject"
...
security:
  - bearerAuth: []